+++ TDEngine2/tools/MeshConverter/meshConverter.h
@@ -38,10 +38,12 @@ namespace TDEngine2
 		std::string mOutputDirname = ".";
 		std::string mOutputFilename = "NewMesh.mesh";
 		std::string mAnimationImportInfoFilename = "AnimationsList.info";
+		std::string mResourcesBuildManifestFilename = "ResourcesBuildSettings.manifest";
 
 		bool mShouldSkipNormals  = false;
 		bool mShouldSkipTangents = false;
 		bool mShouldSkipJoints   = false;
+		bool mIsBucketModeEnabled = false;
 
 		U32 mIndexFormat = sizeof(U16);
 	};